+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # Change Log
 All notable changes to this project will be documented in this file.
+#### 0.9.146 Releases
+
+- `0.9.146` Releases - [0.9.146](#09146)
 
 #### 0.9.137 Releases
 
 - `0.9.137` Releases - [0.9.137](#09137)
 
 ---
+## [0.9.146](https://github.com/ciscospark/spark-ios-sdk/releases/tag/0.9.146)
+Released on 2016-05-19.
+
+#### Added
+- Add CHANGELOG
+- Support refreshing token
+
+#### Updated
+- Refine OAuth flow logic
 
 ## [0.9.137](https://github.com/ciscospark/spark-ios-sdk/releases/tag/0.9.137)
 Released on 2016-05-12.